refactor: client code part 1
