+++ client/src/app/common.rs
@@ -2,6 +2,7 @@ pub(crate) mod table;
 pub(crate) mod types;
 
 use egui_extras::{TableBody, TableRow};
+
 pub use table::*;
 pub use types::*;
