--- client/src/app/mpegts_streams_table/types.rs
@@ -1,7 +1,0 @@
-use crate::app::common::types::TableConfig;
-use crate::streams::mpegts_stream::substream::MpegtsSubStream;
-
-pub struct StreamInfo {
-    pub substream: MpegtsSubStream,
-    pub id: String,
-}
