+++ client/src/app/tables.rs
@@ -1,0 +1,17 @@
+pub mod mpegts_info_table;
+pub mod mpegts_packets_table;
+pub mod mpegts_streams_table;
+pub mod packets_table;
+pub mod rtcp_packets_table;
+pub mod rtp_packets_table;
+pub mod rtp_streams_plot;
+pub mod rtp_streams_table;
+
+pub use mpegts_info_table::*;
+pub use mpegts_packets_table::*;
+pub use mpegts_streams_table::*;
+pub use packets_table::*;
+pub use rtcp_packets_table::*;
+pub use rtp_packets_table::*;
+pub use rtp_streams_plot::*;
+pub use rtp_streams_table::*;
