+++ client/src/app/tables/mpegts_packets_table/types.rs
@@ -1,5 +1,6 @@
-use crate::app::common::types::TableConfig;
-use crate::streams::mpegts_stream::packet_info::MpegTsPacketInfo;
+use crate::{
+    app::common::types::TableConfig, streams::mpegts_stream::packet_info::MpegTsPacketInfo,
+};
 use netpix_common::MpegtsStreamKey;
 use web_time::Duration;
