+++ client/src/app/tables/mpegts_streams_table.rs
