+++ client/src/app/tables/mpegts_streams_table/display.rs
