+++ client/src/app/tables/mpegts_streams_table/types.rs
@@ -1,0 +1,6 @@
+use crate::{app::common::types::TableConfig, streams::mpegts_stream::substream::MpegtsSubStream};
+
+pub struct StreamInfo {
+    pub substream: MpegtsSubStream,
+    pub id: String,
+}
