+++ client/src/app/tables/packets_table/filters.rs
@@ -43,10 +43,12 @@
 //! - `(type:rtp OR type:rtcp) AND NOT dest:10.0.0.1` - RTP/RTCP packets not going to specific host
 //! - `proto:tcp AND length:>=1500` - TCP packets with maximum size
 
-use crate::filter_system::{
-    CommonFilterParser, ComparisonFilter, FilterExpression, FilterParser, ParseError,
+use crate::{
+    declare_filter_type,
+    filter_system::{
+        self, CommonFilterParser, ComparisonFilter, FilterExpression, FilterParser, ParseError,
+    },
 };
-use crate::{declare_filter_type, filter_system};
 use netpix_common::packet::{Packet, SessionProtocol};
 use std::str::FromStr;
