+++ client/src/app/tables/packets_table/types.rs
