+++ client/src/app/tables/rtcp_packets_table.rs
@@ -1,9 +1,7 @@
-use crate::streams::RefStreams;
-use crate::utils::ntp_to_string;
+use crate::{streams::RefStreams, utils::ntp_to_string};
 use egui::{RichText, Ui};
 use egui_extras::{Column, TableBody, TableBuilder};
-use netpix_common::rtcp::*;
-use netpix_common::{packet::SessionPacket, RtcpPacket};
+use netpix_common::{packet::SessionPacket, rtcp::*, RtcpPacket};
 
 pub struct RtcpPacketsTable {
     streams: RefStreams,
