+++ client/src/streams/mpegts_stream.rs
@@ -1,30 +1,36 @@
 #![allow(dead_code)]
-use crate::streams::mpegts_stream::substream::{
-    MpegtsSubStream, MpegtsSubStreams, SubStreamKey, SubstreamMpegTsPacketInfo,
+use crate::streams::{
+    mpegts_stream::substream::{
+        MpegtsSubStream, MpegtsSubStreams, SubStreamKey, SubstreamMpegTsPacketInfo,
+    },
+    stream_statistics::{Bitrate, Bytes, PacketsTime, Statistics, StreamStatistics},
 };
-use crate::streams::stream_statistics::{
-    Bitrate, Bytes, PacketsTime, Statistics, StreamStatistics,
+use netpix_common::{
+    mpegts::{
+        aggregator::MpegtsAggregator,
+        header::{AdaptationFieldControl, PIDTable},
+        psi::{
+            pat::{fragmentary_pat::FragmentaryProgramAssociationTable, ProgramAssociationTable},
+            pmt::{fragmentary_pmt::FragmentaryProgramMapTable, ProgramMapTable},
+            psi_buffer::{FragmentaryPsi, PsiBuffer},
+        },
+        MpegtsFragment,
+    },
+    packet::SessionPacket,
+    MpegtsPacket, Packet, PacketAssociationTable,
 };
-use netpix_common::mpegts::aggregator::MpegtsAggregator;
-use netpix_common::mpegts::header::{AdaptationFieldControl, PIDTable};
-use netpix_common::mpegts::psi::pat::fragmentary_pat::FragmentaryProgramAssociationTable;
-use netpix_common::mpegts::psi::pat::ProgramAssociationTable;
-use netpix_common::mpegts::psi::pmt::fragmentary_pmt::FragmentaryProgramMapTable;
-use netpix_common::mpegts::psi::pmt::ProgramMapTable;
-use netpix_common::mpegts::psi::psi_buffer::{FragmentaryPsi, PsiBuffer};
-use netpix_common::mpegts::MpegtsFragment;
-use netpix_common::packet::SessionPacket;
-use netpix_common::{MpegtsPacket, Packet, PacketAssociationTable};
 use rustc_hash::FxHashMap;
-use std::cmp::{max, min};
-use std::time::Duration;
+use std::{
+    cmp::{max, min},
+    time::Duration,
+};
+
+use packet_info::{MpegTsPacketInfo, MpegTsStreamInfo};
+use packet_processor::MpegtsPacketProcessor;
 
 pub mod packet_info;
 pub mod packet_processor;
 pub mod substream;
-
-use packet_info::{MpegTsPacketInfo, MpegTsStreamInfo};
-use packet_processor::MpegtsPacketProcessor;
 
 #[derive(Debug, Clone)]
 pub struct MpegTsStream {
