+++ client/src/streams/mpegts_stream/packet_info.rs
@@ -1,8 +1,11 @@
 use crate::streams::stream_statistics::{Bitrate, Bytes, PacketsTime, Statistics};
-use netpix_common::mpegts::header::{AdaptationFieldControl, PIDTable};
-use netpix_common::mpegts::psi::pat::ProgramAssociationTable;
-use netpix_common::mpegts::psi::pmt::ProgramMapTable;
-use netpix_common::{MpegtsPacket, Packet, PacketAssociationTable};
+use netpix_common::{
+    mpegts::{
+        header::{AdaptationFieldControl, PIDTable},
+        psi::{pat::ProgramAssociationTable, pmt::ProgramMapTable},
+    },
+    MpegtsPacket, Packet, PacketAssociationTable,
+};
 use rustc_hash::FxHashMap;
 use std::time::Duration;
