+++ client/src/streams/mpegts_stream/packet_processor.rs
@@ -1,17 +1,23 @@
-use crate::streams::mpegts_stream::packet_info::{MpegTsPacketInfo, MpegTsStreamInfo};
-use crate::streams::mpegts_stream::substream::{
-    MpegtsSubStream, MpegtsSubStreams, SubStreamParameters, SubstreamMpegTsPacketInfo,
+use crate::streams::mpegts_stream::{
+    packet_info::{MpegTsPacketInfo, MpegTsStreamInfo},
+    substream::{
+        MpegtsSubStream, MpegtsSubStreams, SubStreamParameters, SubstreamMpegTsPacketInfo,
+    },
 };
-use netpix_common::mpegts::aggregator::MpegtsAggregator;
-use netpix_common::mpegts::header::PIDTable;
-use netpix_common::mpegts::psi::pat::fragmentary_pat::FragmentaryProgramAssociationTable;
-use netpix_common::mpegts::psi::pat::ProgramAssociationTable;
-use netpix_common::mpegts::psi::pmt::fragmentary_pmt::FragmentaryProgramMapTable;
-use netpix_common::mpegts::psi::pmt::ProgramMapTable;
-use netpix_common::mpegts::psi::psi_buffer::{FragmentaryPsi, PsiBuffer};
-use netpix_common::mpegts::MpegtsFragment;
-use netpix_common::packet::SessionPacket;
-use netpix_common::{MpegtsPacket, Packet, PacketAssociationTable};
+use netpix_common::{
+    mpegts::{
+        aggregator::MpegtsAggregator,
+        header::PIDTable,
+        psi::{
+            pat::{fragmentary_pat::FragmentaryProgramAssociationTable, ProgramAssociationTable},
+            pmt::{fragmentary_pmt::FragmentaryProgramMapTable, ProgramMapTable},
+            psi_buffer::{FragmentaryPsi, PsiBuffer},
+        },
+        MpegtsFragment,
+    },
+    packet::SessionPacket,
+    MpegtsPacket, Packet, PacketAssociationTable,
+};
 use std::time::Duration;
 
 #[derive(Debug, Clone)]
